[Player] Extract CurrencyMgr from Player.cpp
Third Player.cpp god-object extraction in the 2026-05-12 campaign,
following GlyphMgr (d9eeec78d) and HonorMgr (e4265c1b5). Moves the
Cata-era unified currency system (honor points, conquest points,
justice, valor, plus per-meta categories) into a new CurrencyMgr
class. By line count this is the largest of the three; by API
surface it is the widest (14 methods, struct + 2 enums + typedef).

What moves into new src/game/Object/CurrencyMgr.{cpp,h}:
  Types previously in Player.h:
    - enum PlayerCurrencyFlag
    - enum PlayerCurrencyState
    - struct PlayerCurrency
    - typedef PlayerCurrenciesMap
  State previously on Player:
    - PlayerCurrenciesMap m_currencies
  Methods previously on Player:
    - GetCurrencyCount          -> CurrencyMgr::GetCount
    - GetCurrencySeasonCount    -> CurrencyMgr::GetSeasonCount
    - GetCurrencyWeekCount      -> CurrencyMgr::GetWeekCount
    - GetCurrencyWeekCap        -> CurrencyMgr::GetWeekCap
    - GetCurrencyTotalCap       -> CurrencyMgr::GetTotalCap
    - ModifyCurrencyCount       -> CurrencyMgr::ModifyCount
    - SetCurrencyCount          -> CurrencyMgr::SetCount
    - SetCurrencyFlags          -> CurrencyMgr::SetFlags
    - SendCurrencies            -> CurrencyMgr::SendAll
    - SendCurrencyWeekCap (x2)  -> CurrencyMgr::SendWeekCap (x2)
    - ResetCurrencyWeekCounts   -> CurrencyMgr::ResetWeekCounts
    - _LoadCurrencies           -> CurrencyMgr::Load
    - _SaveCurrencies           -> CurrencyMgr::Save

What stays on Player.{cpp,h}:
  - Full public Currency API surface (GetCurrencyCount, ModifyCurrency
    Count, etc.) as inline delegates in Player.h. External callers -
    BattleGround::1001, LootHandler:139, HonorMgr::Reward:213,
    PlayerCommands:2577/2579, PlayerHonorCommands:112, ListCommands:
    651, CharacterHandler::HandleSetCurrencyFlagsOpcode:1674 - do NOT
    need to change. The Player API matches the previous signatures
    1:1, including default-argument values.
  - HasCurrencyCount / HasCurrencySeasonCount: trivial inline
    convenience methods on Player that call the count getters.
  - BuyCurrencyFromVendorSlot: vendor-system integration that uses
    ModifyCurrencyCount; stays on Player.
  - CurrencyAddedQuestCheck / CurrencyRemovedQuestCheck: quest-tracker
    side-effects of currency mutation; CurrencyMgr::ModifyCount calls
    them through m_owner so the quest hooks still fire at the right
    moment.

Net change:
  - Player.cpp: -357 lines (the 14 function bodies)
  - Player.h:    -65 net (types removed + member replaced + 14 decls
    converted to inline delegates)
  - CurrencyMgr.{h,cpp}: ~480 lines combined of relocated code.

Cumulative effect of the three-extraction campaign:
  - Player.cpp: 28,617 -> ~27,930 lines (-687 net so far across
    Glyph + Honor + Currency).
  - Three independent managers now own their respective state and
    lifecycle; Player retains exactly the public API surface it had
    pre-campaign.

Smoke-tested on 2026-05-12 (single-tester):
  - Character load: Player::LoadFromDB now calls
    m_currencyMgr.Load(result) via delegate. Loads clean, the
    Currency tab UI populates with the expected balances.
  - `.modify currency 392 5000` (admin command) correctly added
    50 displayed Honor Points (internal 5000 / precision 100 = 50).
    Lower values incrementally accumulate as expected against the
    precision divisor - confirms CurrencyMgr::ModifyCount mutates
    state, SMSG_SET_CURRENCY emits the precision-divided value, and
    the client live-updates the UI.
  - Log out and log back in: balances persisted correctly. Confirms
    CurrencyMgr::Save -> character_currencies INSERT/UPDATE/DELETE
    branches and CurrencyMgr::Load round-trip both work via the
    Player delegates from SaveToDB / LoadFromDB.
