+++ src/game/Object/Player.cpp
@@ -311,7 +311,7 @@ UpdateMask Player::updateVisualBits;
  *
  * @param session The owning world session.
  */
-Player::Player(WorldSession* session): Unit(), m_mover(this), m_camera(this), m_petMgr(this), m_achievementMgr(this), m_reputationMgr(this), m_glyphMgr(this), m_honorMgr(this)
+Player::Player(WorldSession* session): Unit(), m_mover(this), m_camera(this), m_petMgr(this), m_achievementMgr(this), m_reputationMgr(this), m_glyphMgr(this), m_honorMgr(this), m_currencyMgr(this)
 {
     m_transport = 0;
 
@@ -29910,357 +29910,12 @@ void Player::_fillGearScoreData(Item* item, GearScoreVec* gearScore, uint32& two
     }
 }
 
-void Player::SendCurrencies() const
-{
-    WorldPacket data(SMSG_SEND_CURRENCIES, m_currencies.size() * 4);
-    data.WriteBits(m_currencies.size(), 23);
+// Player::SendCurrencies / Get/SendCurrencyWeekCap / GetCurrencyTotalCap / Get*Count moved
+// to CurrencyMgr (2026-05-12); thin delegating wrappers live inline in Player.h.
 
-    for (PlayerCurrenciesMap::const_iterator itr = m_currencies.begin(); itr != m_currencies.end(); ++itr)
-    {
-        uint32 weekCap = GetCurrencyWeekCap(itr->second.currencyEntry);
-        data.WriteBit(weekCap && itr->second.weekCount);
-        data.WriteBits(itr->second.flags, 4);
-        data.WriteBit(weekCap);
-        data.WriteBit(itr->second.currencyEntry->HasSeasonCount());
-    }
-
-    for (PlayerCurrenciesMap::const_iterator itr = m_currencies.begin(); itr != m_currencies.end(); ++itr)
-    {
-        data << uint32(floor(itr->second.totalCount / itr->second.currencyEntry->GetPrecision()));
-
-        uint32 weekCap = GetCurrencyWeekCap(itr->second.currencyEntry);
-        if (weekCap)
-        {
-            data << uint32(floor(weekCap / itr->second.currencyEntry->GetPrecision()));
-        }
-        if (itr->second.currencyEntry->HasSeasonCount())
-        {
-            data << uint32(floor(itr->second.seasonCount / itr->second.currencyEntry->GetPrecision()));
-        }
-        data << uint32(itr->first);
-        if (weekCap && itr->second.weekCount)
-        {
-            data << uint32(floor(itr->second.weekCount / itr->second.currencyEntry->GetPrecision()));
-        }
-    }
-
-    GetSession()->SendPacket(&data);
-}
-
-uint32 Player::GetCurrencyWeekCap(CurrencyTypesEntry const * currency) const
-{
-    uint32 cap = currency->WeekCap;
-    switch (currency->ID)
-    {
-        case CURRENCY_CONQUEST_POINTS:
-            cap = sWorld.getConfig(CONFIG_UINT32_CURRENCY_CONQUEST_POINTS_DEFAULT_WEEK_CAP);
-            break;
-    }
-
-    return cap;
-}
-
-void Player::SendCurrencyWeekCap(uint32 id) const
-{
-    SendCurrencyWeekCap(sCurrencyTypesStore.LookupEntry(id));
-}
-
-void Player::SendCurrencyWeekCap(CurrencyTypesEntry const * currency) const
-{
-    if (!currency || !IsInWorld() || GetSession()->PlayerLoading())
-    {
-        return;
-    }
-
-    uint32 cap = GetCurrencyWeekCap(currency);
-    if (!cap)
-    {
-        return;
-    }
-
-    WorldPacket packet(SMSG_SET_CURRENCY_WEEK_LIMIT, 8);
-    packet << uint32(floor(cap / currency->GetPrecision()));
-    packet << uint32(currency->ID);
-    GetSession()->SendPacket(&packet);
-}
-
-uint32 Player::GetCurrencyTotalCap(CurrencyTypesEntry const* currency) const
-{
-    uint32 cap = currency->TotalCap;
-
-    return cap;
-}
-
-uint32 Player::GetCurrencyCount(uint32 id) const
-{
-    PlayerCurrenciesMap::const_iterator itr = m_currencies.find(id);
-    return itr != m_currencies.end() ? itr->second.totalCount : 0;
-}
-
-uint32 Player::GetCurrencySeasonCount(uint32 id) const
-{
-    PlayerCurrenciesMap::const_iterator itr = m_currencies.find(id);
-    return itr != m_currencies.end() ? itr->second.seasonCount : 0;
-}
-
-uint32 Player::GetCurrencyWeekCount(uint32 id) const
-{
-    PlayerCurrenciesMap::const_iterator itr = m_currencies.find(id);
-    return itr != m_currencies.end() ? itr->second.weekCount : 0;
-}
-
-void Player::ModifyCurrencyCount(uint32 id, int32 count, bool modifyWeek, bool modifySeason, bool ignoreMultipliers)
-{
-    if (!count)
-    {
-        return;
-    }
-
-    if (!ignoreMultipliers && count > 0)
-    {
-        count *= GetTotalAuraMultiplierByMiscValue(SPELL_AURA_MOD_CURRENCY_GAIN, id);
-    }
-
-    CurrencyTypesEntry const * currency = NULL;
-
-    int32 oldTotalCount = 0;
-    int32 oldWeekCount = 0;
-    PlayerCurrenciesMap::iterator itr = m_currencies.find(id);
-
-    bool initWeek = false;
-    if (itr == m_currencies.end())
-    {
-        currency = sCurrencyTypesStore.LookupEntry(id);
-        MANGOS_ASSERT(currency);
-
-        PlayerCurrency cur;
-        cur.state = PLAYERCURRENCY_NEW;
-        cur.totalCount = 0;
-        cur.weekCount = 0;
-        cur.seasonCount = 0;
-        cur.flags = 0;
-        cur.currencyEntry = currency;
-        m_currencies[id] = cur;
-        initWeek = true;
-        itr = m_currencies.find(id);
-    }
-    else
-    {
-        oldTotalCount = itr->second.totalCount;
-        oldWeekCount = itr->second.weekCount;
-        currency = itr->second.currencyEntry;
-    }
-
-    int32 newTotalCount = oldTotalCount + count;
-    if (newTotalCount < 0)
-    {
-        newTotalCount = 0;
-    }
-
-    int32 newWeekCount = oldWeekCount + (modifyWeek && count > 0 ? count : 0);
-    if (newWeekCount < 0)
-    {
-        newWeekCount = 0;
-    }
-
-    int32 totalCap = GetCurrencyTotalCap(currency);
-    if (totalCap && totalCap < newTotalCount)
-    {
-        int32 delta = newTotalCount - totalCap;
-        newTotalCount = totalCap;
-        newWeekCount -= delta;
-    }
-
-    int32 weekCap = GetCurrencyWeekCap(currency);
-    if (modifyWeek && weekCap && newWeekCount > weekCap)
-    {
-        int32 delta = newWeekCount - weekCap;
-        newWeekCount = weekCap;
-        newTotalCount -= delta;
-    }
-    initWeek &= weekCap != currency->WeekCap;
-
-    if (newTotalCount != oldTotalCount)
-    {
-        if (itr->second.state != PLAYERCURRENCY_NEW)
-        {
-            itr->second.state = PLAYERCURRENCY_CHANGED;
-        }
-
-        itr->second.totalCount = newTotalCount;
-        itr->second.weekCount = newWeekCount;
-
-        int32 diff = newTotalCount - oldTotalCount;
-        if (diff > 0 && modifySeason)
-        {
-            itr->second.seasonCount += diff;
-        }
-
-        // probably excessive checks
-        if (IsInWorld() && !GetSession()->PlayerLoading())
-        {
-            if (diff > 0)
-            {
-                UpdateAchievementCriteria(ACHIEVEMENT_CRITERIA_TYPE_CURRENCY_EARNED, id, newTotalCount);
-            }
-
-            WorldPacket packet(SMSG_SET_CURRENCY, 13);
-            bool bit0 = modifyWeek && weekCap && diff > 0;
-            bool bit1 = currency->HasSeasonCount();
-            bool bit2 = currency->Category == CURRENCY_CATEGORY_META;   // hides message in client when set
-            packet.WriteBit(bit0);
-            packet.WriteBit(bit1);
-            packet.WriteBit(bit2);
-
-            if (bit1)
-            {
-                packet << uint32(floor(itr->second.seasonCount / currency->GetPrecision()));
-            }
-            packet << uint32(floor(newTotalCount / currency->GetPrecision()));
-            packet << uint32(id);
-            if (bit0)
-            {
-                packet << uint32(floor(newWeekCount / currency->GetPrecision()));
-            }
-            GetSession()->SendPacket(&packet);
-
-            // init currency week limit for new currencies
-            if (initWeek)
-            {
-                SendCurrencyWeekCap(currency);
-            }
-
-            if (diff > 0)
-            {
-                CurrencyAddedQuestCheck(id);
-            }
-            else
-            {
-                CurrencyRemovedQuestCheck(id);
-            }
-        }
-
-        if (itr->first == CURRENCY_CONQUEST_ARENA_META || itr->first == CURRENCY_CONQUEST_BG_META)
-        {
-            ModifyCurrencyCount(CURRENCY_CONQUEST_POINTS, diff, modifyWeek, modifySeason, ignoreMultipliers);
-        }
-    }
-}
-
-void Player::SetCurrencyCount(uint32 id, uint32 count)
-{
-    ModifyCurrencyCount(id, int32(count) - GetCurrencyCount(id), false, false, true);
-}
-
-void Player::_LoadCurrencies(QueryResult* result)
-{
-    //         0   1           2          4            5
-    // "SELECT id, totalCount, weekCount, seasonCount, flags FROM character_currencies WHERE guid = '%u'"
-
-    if (result)
-    {
-        do
-        {
-            Field* fields = result->Fetch();
-
-            uint32 currency_id = fields[0].GetUInt16();
-            uint32 totalCount  = fields[1].GetUInt32();
-            uint32 weekCount   = fields[2].GetUInt32();
-            uint32 seasonCount = fields[3].GetUInt32();
-            uint8 flags        = fields[4].GetUInt8();
-
-            CurrencyTypesEntry const * entry = sCurrencyTypesStore.LookupEntry(currency_id);
-            if (!entry)
-            {
-                sLog.outError("Player::_LoadCurrencies: %s has not existing currency id %u, removing.", GetGuidStr().c_str(), currency_id);
-                CharacterDatabase.PExecute("DELETE FROM `character_currencies` WHERE `id` = '%u'", currency_id);
-                continue;
-            }
-
-            uint32 weekCap = GetCurrencyWeekCap(entry);
-            uint32 totalCap = GetCurrencyTotalCap(entry);
-
-            PlayerCurrency cur;
-
-            cur.state = PLAYERCURRENCY_UNCHANGED;
-
-            if (totalCap && totalCount > totalCap)
-            {
-                cur.totalCount = totalCap;
-            }
-            else
-            {
-                cur.totalCount = totalCount;
-            }
-
-            if (weekCap && weekCount > weekCap)
-            {
-                cur.weekCount = weekCap;
-            }
-            else
-            {
-                cur.weekCount = weekCount;
-            }
-
-            cur.seasonCount = seasonCount;
-
-            cur.flags = flags & PLAYERCURRENCY_MASK_USED_BY_CLIENT;
-            cur.currencyEntry = entry;
-
-            m_currencies[currency_id] = cur;
-        }
-        while (result->NextRow());
-    }
-}
-
-void Player::_SaveCurrencies()
-{
-    for (PlayerCurrenciesMap::iterator itr = m_currencies.begin(); itr != m_currencies.end();)
-    {
-        if (itr->second.state == PLAYERCURRENCY_CHANGED)
-        {
-            CharacterDatabase.PExecute("UPDATE `character_currencies` SET `totalCount` = '%u', `weekCount` = '%u', `seasonCount` = '%u', `flags` = '%u' WHERE `guid` = '%u' AND `id` = '%u'", itr->second.totalCount, itr->second.weekCount, itr->second.seasonCount, itr->second.flags, GetGUIDLow(), itr->first);
-        }
-        else if (itr->second.state == PLAYERCURRENCY_NEW)
-        {
-            CharacterDatabase.PExecute("INSERT INTO `character_currencies` (`guid`, `id`, `totalCount`, `weekCount`, `seasonCount`, `flags`) VALUES ('%u', '%u', '%u', '%u', '%u', '%u')", GetGUIDLow(), itr->first, itr->second.totalCount, itr->second.weekCount, itr->second.seasonCount, itr->second.flags);
-        }
-
-        if (itr->second.state == PLAYERCURRENCY_REMOVED)
-        {
-            m_currencies.erase(itr++);
-        }
-        else
-        {
-            itr->second.state = PLAYERCURRENCY_UNCHANGED;
-            ++itr;
-        }
-    }
-}
-
-void Player::SetCurrencyFlags(uint32 currencyId, uint8 flags)
-{
-    PlayerCurrenciesMap::iterator itr = m_currencies.find(currencyId);
-    if (itr == m_currencies.end())
-    {
-        return;
-    }
-
-    itr->second.flags = flags;
-    itr->second.state = PLAYERCURRENCY_CHANGED;
-}
-
-void Player::ResetCurrencyWeekCounts()
-{
-    for (PlayerCurrenciesMap::iterator itr = m_currencies.begin(); itr != m_currencies.end(); ++itr)
-    {
-        itr->second.weekCount = 0;
-        itr->second.state = PLAYERCURRENCY_CHANGED;
-    }
-
-    WorldPacket data(SMSG_WEEKLY_RESET_CURRENCIES, 0);
-    SendDirectMessage(&data);
-}
+// Player::ModifyCurrencyCount / SetCurrencyCount / _LoadCurrencies / _SaveCurrencies /
+// SetCurrencyFlags / ResetCurrencyWeekCounts moved to CurrencyMgr (2026-05-12);
+// thin delegating wrappers live inline in Player.h.
 
 void Player::SendPvPRewards()
 {
